# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Campus College Management API
  version: 1.0.0
  description: Versioned REST API for role-scoped college operations.
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
security:
  - sessionCookie: []
paths:
...
  /auth/login:
    post:
      security: []
      summary: Create an authenticated session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { description: Signed in; session cookie is set }
        "401": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: college_session
...
  schemas:
...
    Error:
      type: object
      properties:
        success: { type: boolean, const: false }
        message: { type: string }
        errors:
          type: array
          items: { type: object }
  responses:
    Error:
      description: Safe API error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"